fix(types): fix Any import regression, float severity metrics and test-list/path annotations (-6)

+++ src/evaluation/metrics.py
@@ -120,7 +120,7 @@ class EvaluationMetrics:
     model_name: str = ""
 
     # Per-severity breakdown
-    by_severity: Dict[str, Dict[str, int]] = field(default_factory=dict)
+    by_severity: Dict[str, Dict[str, float]] = field(default_factory=dict)
 
     def to_dict(self) -> Dict[str, Any]:
         """Convert metrics to dictionary."""
